+++ src/main/java/io/github/rciam/keycloak/resolver/ThemeConfig.java
@@ -3,13 +3,7 @@ package io.github.rciam.keycloak.resolver;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.github.rciam.keycloak.resolver.stubs.Configuration;
-import io.github.rciam.keycloak.resolver.stubs.cluster.RealmCreatedEvent;
-import io.github.rciam.keycloak.resolver.stubs.cluster.RealmDeletedEvent;
 import org.jboss.logging.Logger;
-import org.keycloak.cluster.ClusterEvent;
-import org.keycloak.cluster.ClusterProvider;
-import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.RealmModel;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
@@ -34,19 +28,15 @@ public class ThemeConfig {
 
     private static final Logger logger = Logger.getLogger(ThemeConfig.class);
 
-    private static boolean REALMS_LISTENER_ADDED = false;
-    private static final String CREATE_THEME_CONFIG = "CREATE_THEME_CONFIG";
-    private static final String DELETE_THEME_CONFIG = "DELETE_THEME_CONFIG";
-    private static ClusterProvider cluster;
     private static WatchService watchService;
     private static WatchKey watchKey;
     private static boolean FOLDER_INITIALIZED = false;
     private static Configuration defaultConfig;
-    private static Map<String, Configuration> realmsConfigs;
+    private static Map<String, Configuration> realmsConfigs = new HashMap<>();
 
 
-    public ThemeConfig(KeycloakSession session) {
-        initializeStatics(session);
+    public ThemeConfig() {
+        initializeStatics();
     }
 
 
@@ -72,7 +62,7 @@ public class ThemeConfig {
     }
 
 
-    private void initializeStatics(KeycloakSession session) {
+    private void initializeStatics() {
 
         if(!FOLDER_INITIALIZED) {
             try {
@@ -85,47 +75,6 @@ public class ThemeConfig {
 
         if(defaultConfig == null)
             loadDefaultConfig();
-
-        if(realmsConfigs == null){
-            realmsConfigs = new HashMap<>();
-            session.realms().getRealmsStream().forEach(realmModel -> {
-                localSynchronizeConfig(realmModel.getName());
-            });
-        }
-
-        if(cluster == null)
-            cluster = session.getProvider(ClusterProvider.class);
-
-        if(!REALMS_LISTENER_ADDED){
-            //register local listener
-            session.getKeycloakSessionFactory().register(event -> {
-                if(event instanceof RealmModel.RealmCreationEvent) {
-                    String realmName = ((RealmModel.RealmCreationEvent)event).getCreatedRealm().getName();
-                    localSynchronizeConfig(realmName);
-                    cluster.notify(CREATE_THEME_CONFIG, RealmCreatedEvent.create(realmName), true); //broadcast event to all other cluster nodes
-                }
-                else if(event instanceof RealmModel.RealmRemovedEvent) {
-                    String realmName = ((RealmModel.RealmRemovedEvent)event).getRealm().getName();
-                    String filePath = getThemeConfigFilePath(realmName);
-                    File configFile = new File(filePath);
-                    if(configFile.exists())
-                        configFile.delete();
-                    cluster.notify(DELETE_THEME_CONFIG, RealmDeletedEvent.create(realmName), true); //broadcast event to all other cluster nodes
-                }
-            });
-            //register cluster listeners
-            cluster.registerListener(CREATE_THEME_CONFIG, (ClusterEvent event) -> {
-                RealmCreatedEvent realmCreatedEvent = (RealmCreatedEvent) event;
-                localSynchronizeConfig(realmCreatedEvent.getRealmName());
-            });
-            cluster.registerListener(DELETE_THEME_CONFIG, (ClusterEvent event) -> {
-                RealmDeletedEvent realmDeletedEvent = (RealmDeletedEvent) event;
-                localSynchronizeConfig(realmDeletedEvent.getRealmName());
-            });
-
-            REALMS_LISTENER_ADDED = true;
-        }
-
 
         if(watchKey == null) {
             Runnable runnable = () -> {
@@ -180,7 +129,7 @@ public class ThemeConfig {
     }
 
 
-    private void localSynchronizeConfig(String realmName) {
+    public void localSynchronizeConfig(String realmName) {
         localSynchronizeConfig(realmName, null);
     }
 
@@ -232,7 +181,7 @@ public class ThemeConfig {
         return String.format("%s/%s/%s", Commons.getBasePath(), Commons.THEME_WORKING_FOLDER, Commons.CONFIGURATIONS_FOLDER);
     }
 
-    private String getThemeConfigFilePath(String realmName){
+    public String getThemeConfigFilePath(String realmName){
         return String.format("%s/%s/%s/%s.json", Commons.getBasePath(), Commons.THEME_WORKING_FOLDER, Commons.CONFIGURATIONS_FOLDER, realmName);
     }